Updated layout
Updated the layout of the project links so they all fit on the screen. Removed the copyright at the bottom as it was interfering with the links (and I don't have one so I have no idea why it was there). Also shrunk the video player sizes so they can fit on smaller screens easier.